Move Android `StringValue` into a new module
